core: resolve InclChunk parser faults, improve JB2 include ...
shared image dictionaries, fix logical error in DirmComponent
properties decoding, fix PixelMap pixel operations using byte,
add DjvuNet.DjvuLibre.Compatibility.Tests project aimed at
direct testing of binary compatibility between DjvuNet and
libdjvulibre.

1. Parser Reliability & GitHub Issue #148: Fixed a critical
NullReferenceException in SjbzChunk. The InclChunk parser was
updated to emulate DjVuLibre's behavior by explicitly byte-walking
and stripping whitespace (\n, \r) from string dictionaries.
Re-architected the Sjbz shared dictionary resolution to robustly
match across ID, Name, and Title. Corrected bitwise flag decoding
in DirmComponent.

2. Test Execution Architecture: Implemented the core engine support for
the DjvuDocFixture caching proposal and to enable final and intermediate
image caching by decoded documents added conditional image caching
to DjvuImage. This behaviour is controlled by an internal IsTesting flag
set in DjvuDocument and DjvuNetImage. When active, DjvuImage performs
aggressive, lazy-initialized caching of intermediate decoding layers
(Background, Foreground, Mask) in addition to existing caching of final
composite images and thumbnails.

3. Binary Compatibility Pipeline: Added native API to libdjvulibre
exposing internal states needed for testing of DIRM, INCL and Sjbz
chunks, corresponding P/Invoke bindings expose them via DjvuNet.DjvuLibre
for compatibility testing (e.g., ddjvu_page_get_jb2_blit).
This allows direct binary comparison of internal C++ state required
for ensuring binary compatibility with libdjvulibre.

4. Bugtrack Fixes: Re-wrote BWT BlockSort tests to accurately reproduce
deep suffix collision faults by restricting overflow padding and at the
same time fixed tests which needed that padding. Fixed hanging BzzWriter
tests by correctly signaling EOF via .Close() to flush the arithmetic
entropy coder. Replaced BSOutputStream loop with BlockCopy.

+++ DjvuNet/Graphics/PixelMap.cs
@@ -1446,15 +1446,15 @@ namespace DjvuNet.Graphics
                         {
                             if (srcpix >= maxgray)
                             {
-                                dst.SetBGR(gTableLocation[fgx.Blue], gTableLocation[fgx.Green],
-                                           gTableLocation[fgx.Red]);
+                                dst.SetBGR(gTableLocation[(byte)fgx.Blue], gTableLocation[(byte)fgx.Green],
+                                           gTableLocation[(byte)fgx.Red]);
                             }
                             else
                             {
                                 int level = multiplier[srcpix];
-                                dst.SetBGR(((dst.Blue * (0x10000 - level)) + (level * gTableLocation[fgx.Blue])) >> 16,
-                                           ((dst.Green * (0x10000 - level)) + (level * gTableLocation[fgx.Green])) >> 16,
-                                           ((dst.Red * (0x10000 - level)) + (level * gTableLocation[fgx.Red])) >> 16);
+                                dst.SetBGR((((byte)dst.Blue * (0x10000 - level)) + (level * gTableLocation[(byte)fgx.Blue])) >> 16,
+                                           (((byte)dst.Green * (0x10000 - level)) + (level * gTableLocation[(byte)fgx.Green])) >> 16,
+                                           (((byte)dst.Red * (0x10000 - level)) + (level * gTableLocation[(byte)fgx.Red])) >> 16);
                             }
                         }
                     }
